Split SVM callback to its own crate
