+++ Cargo.toml
@@ -125,6 +125,7 @@ members = [
     "storage-proto",
     "streamer",
     "svm",
+    "svm-callback",
     "svm-conformance",
     "svm-rent-collector",
     "svm-transaction",
@@ -535,6 +536,7 @@ solana-storage-bigtable = { path = "storage-bigtable", version = "=2.3.0" }
 solana-storage-proto = { path = "storage-proto", version = "=2.3.0" }
 solana-streamer = { path = "streamer", version = "=2.3.0" }
 solana-svm = { path = "svm", version = "=2.3.0" }
+solana-svm-callback = { path = "svm-callback", version = "=2.3.0" }
 solana-svm-conformance = { path = "svm-conformance", version = "=2.3.0" }
 solana-svm-rent-collector = { path = "svm-rent-collector", version = "=2.3.0" }
 solana-svm-transaction = { path = "svm-transaction", version = "=2.3.0" }
