+++ svm/examples/json-rpc/server/Cargo.toml
@@ -29,6 +29,7 @@ solana-program-runtime = { workspace = true }
 solana-rpc-client-api = { workspace = true }
 solana-sdk = { workspace = true }
 solana-svm = { workspace = true }
+solana-svm-callback = { workspace = true }
 solana-system-program = { workspace = true }
 solana-transaction-context = { workspace = true }
 solana-transaction-status = { workspace = true }
