+++ svm/examples/paytube/Cargo.toml
@@ -15,6 +15,7 @@ solana-logger = { workspace = true }
 solana-program-runtime = { workspace = true, features = ["dev-context-only-utils"] }
 solana-sdk = { workspace = true }
 solana-svm = { workspace = true }
+solana-svm-callback = { workspace = true }
 solana-system-program = { workspace = true }
 spl-associated-token-account = { workspace = true }
 spl-token = { workspace = true }
