+++ svm/src/program_loader.rs
@@ -1,5 +1,4 @@
 use {
-    crate::transaction_processing_callback::TransactionProcessingCallback,
     solana_account::{state_traits::StateMut, AccountSharedData, ReadableAccount},
     solana_clock::Slot,
     solana_instruction::error::InstructionError,
@@ -11,6 +10,7 @@ use {
     },
     solana_pubkey::Pubkey,
     solana_sdk_ids::{bpf_loader, bpf_loader_deprecated, bpf_loader_upgradeable, loader_v4},
+    solana_svm_callback::TransactionProcessingCallback,
     solana_timings::ExecuteTimings,
     solana_transaction_error::{TransactionError, TransactionResult},
     solana_type_overrides::sync::Arc,
