+++ svm/src/transaction_processor.rs
@@ -15,7 +15,6 @@ use {
         transaction_account_state_info::TransactionAccountStateInfo,
         transaction_error_metrics::TransactionErrorMetrics,
         transaction_execution_result::{ExecutedTransaction, TransactionExecutionDetails},
-        transaction_processing_callback::TransactionProcessingCallback,
         transaction_processing_result::{ProcessedTransaction, TransactionProcessingResult},
     },
     agave_feature_set::{remove_accounts_executable_flag_checks, FeatureSet},
@@ -48,6 +47,7 @@ use {
     solana_pubkey::Pubkey,
     solana_rent_collector::RentCollector,
     solana_sdk_ids::{native_loader, system_program},
+    solana_svm_callback::TransactionProcessingCallback,
     solana_svm_rent_collector::svm_rent_collector::SVMRentCollector,
     solana_svm_transaction::{svm_message::SVMMessage, svm_transaction::SVMTransaction},
     solana_timings::{ExecuteTimingType, ExecuteTimings},
@@ -1070,7 +1070,6 @@ mod tests {
             account_loader::{LoadedTransactionAccount, ValidatedTransactionDetails},
             nonce_info::NonceInfo,
             rollback_accounts::RollbackAccounts,
-            transaction_processing_callback::AccountState,
         },
         agave_feature_set::FeatureSet,
         agave_reserved_account_keys::ReservedAccountKeys,
@@ -1095,6 +1094,7 @@ mod tests {
         solana_rent_debits::RentDebits,
         solana_sdk_ids::{bpf_loader, system_program, sysvar},
         solana_signature::Signature,
+        solana_svm_callback::AccountState,
         solana_transaction::{sanitized::SanitizedTransaction, Transaction},
         solana_transaction_context::TransactionContext,
         solana_transaction_error::{TransactionError, TransactionError::DuplicateInstruction},
